Add more granular resource sections
If nobody is opposed, I would favor something like this. Otherwise, the shortest requested time is 4 hours which is too long for the dev queue on Sherlock. With this granularity, tiny files can be run on the dev queue. I made it so that "nano" is equivalent to the previous "normal" and "micro" and "milli" max out at 7 days which I think is the longest available on Sherlock.