Generate Summary Plots for ATAC pipeline
This R script can be run as a command line one-liner after running looper summarize and will produce summary plots relevant to ATAC-seq. It is designed to retain consistent formating regardless of the number of samples being analyzed. It is also robust to including pre-alignments or not.  It does introduce some more dependencies that I will detail in the ATAC pipeline readme. There may be multiple ways that this could break so if it breaks on your looper summarize TSV file, send it to me and I'll figure it out.